requirements.txt
streamlit
pandas
sqlalchemy
openpyxl
bcrypt
python-dotenv
